backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.7
openai-agents>=0.1.0
qdrant-client==1.9.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.5.2
